tools: refactor objectstore tool
...to expose the definitions of the object dump
format (RadosDump) for use in rados import/export.

The serialized->RADOS code is now RadosImport, and
the serialized->ObjectStore code is ObjectStoreTool.

This is a step toward #9964, which should use the
same on serialization format as the existing
objectstore-tool dumps.
